refatora a importação do hook useIsMobile para um novo arquivo e remove a implementação duplicada nos componentes
