+++ src/app/hooks/isMobile.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react";
+
+export function useIsMobile() {
+  const [isMobile, setIsMobile] = useState<boolean>(false);
+  useEffect(() => {
+    const MOBILE_BREAKPOINT = 768;
+    const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`);
+    const handleChange = () => {
+      setIsMobile(mql.matches);
+    };
+    setIsMobile(mql.matches);
+    mql.addEventListener("change", handleChange);
+    return () => mql.removeEventListener("change", handleChange);
+  }, []);
+  return isMobile;
+}
