+++ src/app/page.tsx
@@ -11,21 +11,7 @@ import { news } from "./data/noticias";
 import { lectures } from "./data/palestras";
 import { communities } from "./data/comunidades";
 import Link from "next/link";
-
-export function useIsMobile() {
-  const [isMobile, setIsMobile] = useState<boolean>(false);
-  useEffect(() => {
-    const MOBILE_BREAKPOINT = 768;
-    const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`);
-    const handleChange = () => {
-      setIsMobile(mql.matches);
-    };
-    setIsMobile(mql.matches);
-    mql.addEventListener("change", handleChange);
-    return () => mql.removeEventListener("change", handleChange);
-  }, []);
-  return isMobile;
-}
+import { useIsMobile } from "@/app/hooks/isMobile";
 
 const Index = () => {
   const isMobile = useIsMobile();
